fix(frontend): converted observability page into component
