--- agenta-web/src/pages/apps/observability/index.tsx
@@ -1,8 +1,0 @@
-import React from "react"
-import ObservabilityDashboard from "../[app_id]/observability"
-
-const index = () => {
-    return <ObservabilityDashboard />
-}
-
-export default index
