+++ agenta-web/src/services/observability/core/index.ts
@@ -11,7 +11,7 @@ import axios from "@/lib/helpers/axiosConfig"
 
 export const fetchAllTraces = async ({appId, queries}: {appId: string; queries?: string}) => {
     const filterByAppId = appId
-        ? `filtering={"conditions":[{"key":"refs.application_id","operator":"is","value":"${appId}"}]}&`
+        ? `filtering={"conditions":[{"key":"refs.application.id","operator":"is","value":"${appId}"}]}&`
         : ""
 
     const response = await axios.get(
